requirements.txt
fastapi
sqlalchemy
pydantic
uvicorn
pymysql
python-dotenv
requests
cryptography
httpx
google-generativeai
google-genai